Add core(free) breach methods
A "breach" is an instance of a system having been compromised by an attacker and the data disclosed.
For example, Adobe was a breach, Gawker was a breach, etc.

It is possible to return the details of each breach in the system.

`Hibp::Client.new.fetch_breaches` - returns all 387(on commit date) breaches.

Also, it is possible to filter breaches by domain.

`Hibp::Client.new.fetch_breaches('adobe.com')` - Filters the result set to only breaches against the domain specified.
It is possible that one site (and consequently domain), is compromised on multiple occasions.

If just a single breach is required and this can be retrieved by the breach "name".
This is the stable value which may or may not be the same as the breach "title" (which can change).
See the breach model below for more info (https://haveibeenpwned.com/API/v3 The breach model).

`Hibp::Client.new.find_breach('000webhost')` - returns information about requested breach.

A "data class" is an attribute of a record compromised in a breach.
For example, many breaches expose data classes such as "Email addresses" and "Passwords".

`Hibp::Client.new.fetch_data_classes` - returns ordered alphabetically a string array and will
expand over time as new breaches expose previously unseen classes of data.